Выполнил задание #11. Учебный проект: разделяй и властвуй
