+++ src/task.js
@@ -1,9 +1,11 @@
-import {createElement} from './util';
 import moment from 'moment';
+import Component from './component';
 
-export default class Task {
+export default class Task extends Component {
 
   constructor(data) {
+    super();
+
     this._title = data.title;
     this._tags = data.tags;
     this._color = data.color;
@@ -11,72 +13,9 @@ export default class Task {
     this._dueDate = data.dueDate;
     this._repeatingDays = data.repeatingDays;
 
-    this._element = null;
     this._onEdit = null;
 
     this._onEditButtonClick = this._onEditButtonClick.bind(this);
-  }
-
-  _isRepeated() {
-    return Object
-      .values(this._repeatingDays)
-      .some((item) => item === true);
-  }
-
-  _isOverdue() {
-    return moment().unix() > this._dueDate;
-  }
-
-  _getDueDate() {
-    return moment.unix(this._dueDate).format(`DD MMMM`);
-  }
-
-  _getDueTime() {
-    return moment.unix(this._dueDate).format(`LT`);
-  }
-
-  _tagsMarkdown() {
-    return [...this._tags]
-      .map((item) => {
-        return `<span class="card__hashtag-inner">
-              <input type="hidden" name="hashtag" value="repeat" class="card__hashtag-hidden-input">
-              <button type="button" class="card__hashtag-name">
-                #${item}
-              </button>
-              <button type="button" class="card__hashtag-delete">
-                delete
-              </button>
-            </span>
-        `.trim();
-      }).join(``);
-  }
-
-  _onEditButtonClick() {
-    this._onEdit();
-  }
-
-  _bind() {
-    this._element
-      .querySelector(`.card__btn--edit`)
-      .addEventListener(`click`, this._onEditButtonClick);
-  }
-
-  _unbind() {
-    this._element
-      .querySelector(`.card__btn--edit`)
-      .removeEventListener(`click`, this._onEditButtonClick);
-  }
-
-  set onEdit(fn) {
-    if (typeof fn === `function`) {
-      this._onEdit = fn;
-    }
-  }
-
-  get element() {
-    return (this._element)
-      ? this._element
-      : this.render();
   }
 
   get template() {
@@ -142,15 +81,60 @@ export default class Task {
           </article>`.trim();
   }
 
-  render() {
-    this._element = createElement(this.template);
-    this._bind();
-    return this._element;
+  set onEdit(fn) {
+    if (typeof fn === `function`) {
+      this._onEdit = fn;
+    }
   }
 
-  unrender() {
-    this._unbind();
-    this._element = null;
+  _isRepeated() {
+    return Object
+      .values(this._repeatingDays)
+      .some((item) => item === true);
+  }
+
+  _isOverdue() {
+    return moment().unix() > this._dueDate;
+  }
+
+  _getDueDate() {
+    return moment.unix(this._dueDate).format(`DD MMMM`);
+  }
+
+  _getDueTime() {
+    return moment.unix(this._dueDate).format(`LT`);
+  }
+
+  _tagsMarkdown() {
+    return [...this._tags]
+      .map((item) => {
+        return `<span class="card__hashtag-inner">
+              <input type="hidden" name="hashtag" value="repeat" class="card__hashtag-hidden-input">
+              <button type="button" class="card__hashtag-name">
+                #${item}
+              </button>
+              <button type="button" class="card__hashtag-delete">
+                delete
+              </button>
+            </span>
+        `.trim();
+      }).join(``);
+  }
+
+  _bind() {
+    this._element
+      .querySelector(`.card__btn--edit`)
+      .addEventListener(`click`, this._onEditButtonClick);
+  }
+
+  _unbind() {
+    this._element
+      .querySelector(`.card__btn--edit`)
+      .removeEventListener(`click`, this._onEditButtonClick);
+  }
+
+  _onEditButtonClick() {
+    this._onEdit();
   }
 
 }
